requirements.txt
aiogram==3.*
aiosqlite>=0.20.0
python-dotenv>=1.0.0
Pillow>=10.0.0
torch
torchvision
opencv-python
matplotlib
numpy
